LPS-127344 Catch error from callback execution to avoid breaking execution
We don't want to interrupt the execution of the loop, if the execution of the callback has an error, we keep the code running. We won't revalidate because it might not be recoverable due to some internal axe-core error, so we just move on to the next callback in the queue.